+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TablesSidebarPopperController.java
@@ -24,20 +24,8 @@ import java.util.*;
 import static de.mephisto.vpin.ui.Studio.client;
 import static de.mephisto.vpin.ui.util.BindingUtil.debouncer;
 
-public class TablesSidebarPopperController implements Initializable, ChangeListener<Number> {
+public class TablesSidebarPopperController implements Initializable {
   private final static Logger LOG = LoggerFactory.getLogger(TablesSidebarPopperController.class);
-
-  private final static TableStatus STATUS_DISABLED = new TableStatus(0, "InActive (Disabled)");
-  private final static TableStatus STATUS_NORMAL = new TableStatus(1, "Visible (Normal)");
-  private final static TableStatus STATUS_MATURE = new TableStatus(2, "Visible (Mature/Hidden)");
-
-  private final static List<TableStatus> TABLE_STATUSES = Arrays.asList(STATUS_DISABLED, STATUS_NORMAL, STATUS_MATURE);
-
-  @FXML
-  private ComboBox<String> launcherCombo;
-
-  @FXML
-  private ComboBox<TableStatus> statusCombo;
 
   @FXML
   private Button tableEditBtn;
@@ -47,12 +35,6 @@ public class TablesSidebarPopperController implements Initializable, ChangeListe
 
   @FXML
   private SplitMenuButton autoFillBtn;
-
-  @FXML
-  private Button editScreensBtn;
-
-  @FXML
-  private Slider volumeSlider;
 
   @FXML
   private Label emulatorLabel;
@@ -212,86 +194,21 @@ public class TablesSidebarPopperController implements Initializable, ChangeListe
     }
   }
 
-  @FXML
-  private void onScreenEdit() {
-    if (Studio.client.getPinUPPopperService().isPinUPPopperRunning()) {
-      if (Dialogs.openPopperRunningWarning(Studio.stage)) {
-        Dialogs.openPopperScreensDialog(this.game.get());
-        this.refreshView(this.game);
-      }
-      return;
-    }
-
-    Dialogs.openPopperScreensDialog(this.game.get());
-    this.refreshView(this.game);
-  }
-
   @Override
   public void initialize(URL url, ResourceBundle resourceBundle) {
-    statusCombo.setItems(FXCollections.observableList(TABLE_STATUSES));
-    statusCombo.valueProperty().addListener((observableValue, tableStatus, t1) -> {
-      try {
-        if (Studio.client.getPinUPPopperService().isPinUPPopperRunning()) {
-          if (Dialogs.openPopperRunningWarning(Studio.stage)) {
-            this.refreshView(this.game);
-          }
-          return;
-        }
-
-        if (game.isPresent() && tableDetails != null) {
-          if (t1 == null || t1.getValue() == tableDetails.getStatus()) {
-            return;
-          }
-          this.tableDetails.setStatus(t1.getValue());
-          this.tableDetails = client.getPinUPPopperService().saveTableDetails(tableDetails, game.get().getId());
-          EventManager.getInstance().notifyTableChange(this.game.get().getId(), null);
-        }
-      } catch (Exception e) {
-        LOG.error("Failed to save table status: " + e.getMessage(), e);
-        WidgetFactory.showAlert(Studio.stage, "Error", "Failed to table status: " + e.getMessage());
-      }
-    });
-
-    launcherCombo.valueProperty().addListener((observableValue, s, t1) -> {
-      try {
-        if (Studio.client.getPinUPPopperService().isPinUPPopperRunning()) {
-          if (Dialogs.openPopperRunningWarning(Studio.stage)) {
-            this.refreshView(this.game);
-          }
-          return;
-        }
-
-        if (game.isPresent() && tableDetails != null) {
-          String altLauncher = tableDetails.getAltLaunchExe();
-          if (StringUtils.equals(altLauncher, t1)) {
-            return;
-          }
-          this.tableDetails = client.getPinUPPopperService().saveCustomLauncher(this.game.get().getId(), t1);
-        }
-      } catch (Exception e) {
-        LOG.error("Failed to save alt launcher: " + e.getMessage(), e);
-        WidgetFactory.showAlert(Studio.stage, "Error", "Failed to save alt launcher: " + e.getMessage());
-      }
-    });
   }
 
 
   public void setGame(Optional<GameRepresentation> game) {
     this.game = game;
     this.tableDetails = null;
-    this.launcherCombo.setValue(null);
-    this.statusCombo.setValue(null);
     this.refreshView(game);
   }
 
   public void refreshView(Optional<GameRepresentation> g) {
     this.tableEditBtn.setDisable(g.isEmpty());
     this.fixVersionBtn.setDisable(g.isEmpty() || !g.get().isUpdateAvailable());
-    this.editScreensBtn.setDisable(g.isEmpty());
-    volumeSlider.setDisable(g.isEmpty());
     autoFillBtn.setDisable(g.isEmpty());
-    launcherCombo.setDisable(g.isEmpty());
-    statusCombo.setDisable(g.isEmpty());
 
     if (g.isPresent()) {
       GameRepresentation game = g.get();
@@ -304,26 +221,6 @@ public class TablesSidebarPopperController implements Initializable, ChangeListe
       else {
         labelTimesPlayed.setText("0");
       }
-
-
-      List<String> launcherList = new ArrayList<>(tableDetails.getLauncherList());
-      launcherList.add(0, null);
-      launcherCombo.setItems(FXCollections.observableList(launcherList));
-      launcherCombo.setValue(tableDetails.getAltLaunchExe());
-
-      if (tableDetails.getStatus() >= 0 && tableDetails.getStatus() <= 2) {
-        TableStatus tableStatus = TABLE_STATUSES.get(tableDetails.getStatus());
-        statusCombo.setValue(tableStatus);
-      }
-
-      volumeSlider.valueProperty().removeListener(this);
-      if (tableDetails.getVolume() != null) {
-        volumeSlider.setValue(Integer.parseInt(tableDetails.getVolume()));
-      }
-      else {
-        volumeSlider.setValue(100);
-      }
-      volumeSlider.valueProperty().addListener(this);
 
       emulatorLabel.setText(client.getPinUPPopperService().getGameEmulator(tableDetails.getEmulatorId()).getName());
       gameType.setText(tableDetails.getGameType() != null ? tableDetails.getGameType().name() : "-");
@@ -351,9 +248,6 @@ public class TablesSidebarPopperController implements Initializable, ChangeListe
       notes.setText(StringUtils.isEmpty(tableDetails.getNotes()) ? "-" : tableDetails.getNotes());
     }
     else {
-      launcherCombo.setValue(null);
-      volumeSlider.setValue(100);
-
       labelLastPlayed.setText("-");
       labelTimesPlayed.setText("-");
 
@@ -384,68 +278,5 @@ public class TablesSidebarPopperController implements Initializable, ChangeListe
 
   public void setSidebarController(TablesSidebarController tablesSidebarController) {
     this.tablesSidebarController = tablesSidebarController;
-  }
-
-  @Override
-  public void changed(ObservableValue<? extends Number> observable, Number oldValue, Number newValue) {
-    if (game.isPresent()) {
-      final GameRepresentation g = game.get();
-      debouncer.debounce("tableVolume" + g.getId(), () -> {
-        int value = newValue.intValue();
-        if (value == 0) {
-          value = 1;
-        }
-
-        if (tableDetails.getVolume() != null && Integer.parseInt(tableDetails.getVolume()) == value) {
-          return;
-        }
-
-        tableDetails.setVolume(String.valueOf(value));
-        LOG.info("Updates volume of " + g.getGameDisplayName() + " to " + value);
-        try {
-          this.tableDetails = Studio.client.getPinUPPopperService().saveTableDetails(tableDetails, g.getId());
-        } catch (Exception e) {
-          WidgetFactory.showAlert(Studio.stage, e.getMessage());
-        }
-      }, 500);
-    }
-  }
-
-  static class TableStatus {
-    private final int value;
-    private final String label;
-
-    TableStatus(int value, String label) {
-      this.value = value;
-      this.label = label;
-    }
-
-    public int getValue() {
-      return value;
-    }
-
-    public String getLabel() {
-      return label;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-      if (this == o) return true;
-      if (!(o instanceof TableStatus)) return false;
-
-      TableStatus that = (TableStatus) o;
-
-      return value == that.value;
-    }
-
-    @Override
-    public int hashCode() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return label;
-    }
   }
 }