+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/TableDataController.java
@@ -7,23 +7,32 @@ import de.mephisto.vpin.restclient.popper.TableDetails;
 import de.mephisto.vpin.restclient.tables.GameRepresentation;
 import de.mephisto.vpin.ui.Studio;
 import de.mephisto.vpin.ui.events.EventManager;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.*;
 import javafx.stage.Stage;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 import java.util.ResourceBundle;
 
 public class TableDataController implements Initializable, DialogController {
   private final static Logger LOG = LoggerFactory.getLogger(TableDataController.class);
+
+  private final static TableStatus STATUS_DISABLED = new TableStatus(0, "InActive (Disabled)");
+  private final static TableStatus STATUS_NORMAL = new TableStatus(1, "Visible (Normal)");
+  private final static TableStatus STATUS_MATURE = new TableStatus(2, "Visible (Mature/Hidden)");
+  private final static TableStatus STATUS_WIP = new TableStatus(3, "Work In Progress");
+
+  private final static List<TableStatus> TABLE_STATUSES = new ArrayList<>(Arrays.asList(STATUS_DISABLED, STATUS_NORMAL, STATUS_MATURE));
+  private final static List<TableStatus> TABLE_STATUSES_15 = new ArrayList<>(Arrays.asList(STATUS_DISABLED, STATUS_NORMAL, STATUS_MATURE, STATUS_WIP));
 
   @FXML
   private Label titleLabel;
@@ -136,16 +145,91 @@ public class TableDataController implements Initializable, DialogController {
   @FXML
   private Tab extrasTab;
 
+  @FXML
+  private Slider volumeSlider;
+
+  //screens
+  @FXML
+  private CheckBox useEmuDefaultsCheckbox;
+
+  @FXML
+  private CheckBox hideAllCheckbox;
+
+  @FXML
+  private CheckBox topperCheckbox;
+
+  @FXML
+  private CheckBox dmdCheckbox;
+
+  @FXML
+  private CheckBox backglassCheckbox;
+
+  @FXML
+  private CheckBox playfieldCheckbox;
+
+  @FXML
+  private CheckBox musicCheckbox;
+
+  @FXML
+  private CheckBox apronCheckbox;
+
+  @FXML
+  private CheckBox wheelbarCheckbox;
+
+  @FXML
+  private CheckBox loadingCheckbox;
+
+  @FXML
+  private CheckBox otherCheckbox;
+
+  @FXML
+  private CheckBox flyerCheckbox;
+
+  @FXML
+  private CheckBox helpCheckbox;
+
+  @FXML
+  private ComboBox<TableDataController.TableStatus> statusCombo;
+
+  @FXML
+  private ComboBox<String> launcherCombo;
+
+  private List<CheckBox> screenCheckboxes = new ArrayList<>();
   private GameRepresentation game;
-  private TableDetails manifest;
+  private TableDetails tableDetails;
 
   @FXML
   private void onSaveClick(ActionEvent e) {
     Stage stage = (Stage) ((Button) e.getSource()).getScene().getWindow();
-    stage.close();
+
+    String value = "";
+    if (useEmuDefaultsCheckbox.isSelected()) {
+      //nothing, empty value for defaults
+    }
+    else if (hideAllCheckbox.isSelected()) {
+      value = "NONE";
+    }
+    else {
+      List<String> result = new ArrayList<>();
+      if (topperCheckbox.isSelected()) result.add("" + 0);
+      if (dmdCheckbox.isSelected()) result.add("" + 1);
+      if (backglassCheckbox.isSelected()) result.add("" + 2);
+      if (playfieldCheckbox.isSelected()) result.add("" + 3);
+      if (musicCheckbox.isSelected()) result.add("" + 4);
+      if (apronCheckbox.isSelected()) result.add("" + 5);
+      if (wheelbarCheckbox.isSelected()) result.add("" + 6);
+      if (loadingCheckbox.isSelected()) result.add("" + 7);
+      if (otherCheckbox.isSelected()) result.add("" + 8);
+      if (flyerCheckbox.isSelected()) result.add("" + 9);
+      if (helpCheckbox.isSelected()) result.add("" + 10);
+
+      value = String.join(",", result);
+    }
+    tableDetails.setKeepDisplays(value);
 
     try {
-      manifest = Studio.client.getPinUPPopperService().saveTableDetails(this.manifest, game.getId());
+      stage.close();
+      tableDetails = Studio.client.getPinUPPopperService().saveTableDetails(this.tableDetails, game.getId());
       EventManager.getInstance().notifyTableChange(game.getId(), null);
     } catch (Exception ex) {
       LOG.error("Error saving table manifest: " + ex.getMessage(), ex);
@@ -161,7 +245,90 @@ public class TableDataController implements Initializable, DialogController {
 
   @Override
   public void initialize(URL url, ResourceBundle resourceBundle) {
+    //screens
+    screenCheckboxes = Arrays.asList(topperCheckbox, dmdCheckbox, backglassCheckbox, playfieldCheckbox, musicCheckbox,
+      apronCheckbox, wheelbarCheckbox, loadingCheckbox, otherCheckbox, flyerCheckbox, helpCheckbox);
 
+    useEmuDefaultsCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        screenCheckboxes.stream().forEach(check -> check.setSelected(false));
+        hideAllCheckbox.setSelected(false);
+      }
+    });
+
+    hideAllCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        screenCheckboxes.stream().forEach(check -> check.setSelected(false));
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+
+    topperCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    dmdCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    backglassCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    playfieldCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    musicCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    apronCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    wheelbarCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    loadingCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    otherCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    flyerCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
+    helpCheckbox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      if (newValue) {
+        hideAllCheckbox.setSelected(false);
+        useEmuDefaultsCheckbox.setSelected(false);
+      }
+    });
   }
 
   @Override
@@ -173,24 +340,24 @@ public class TableDataController implements Initializable, DialogController {
     this.game = game;
     this.titleLabel.setText("Table Data of '" + game.getGameDisplayName() + "'");
 
-    manifest = Studio.client.getPinUPPopperService().getTableDetails(game.getId());
+    tableDetails = Studio.client.getPinUPPopperService().getTableDetails(game.getId());
 
-    gameName.setText(manifest.getGameName());
-    gameFileName.setText(manifest.getGameFileName());
-    gameFileName.textProperty().addListener((observable, oldValue, newValue) -> manifest.setGameFileName(newValue));
-    gameDisplayName.setText(manifest.getGameDisplayName());
-    gameDisplayName.textProperty().addListener((observable, oldValue, newValue) -> manifest.setGameDisplayName(newValue.trim()));
-    gameTheme.setText(manifest.getGameTheme());
-    gameTheme.textProperty().addListener((observable, oldValue, newValue) -> manifest.setGameTheme(newValue));
-    gameVersion.setText(manifest.getGameVersion());
-    gameVersion.textProperty().addListener((observable, oldValue, newValue) -> manifest.setGameVersion(newValue));
+    gameName.setText(tableDetails.getGameName());
+    gameFileName.setText(tableDetails.getGameFileName());
+    gameFileName.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setGameFileName(newValue));
+    gameDisplayName.setText(tableDetails.getGameDisplayName());
+    gameDisplayName.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setGameDisplayName(newValue.trim()));
+    gameTheme.setText(tableDetails.getGameTheme());
+    gameTheme.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setGameTheme(newValue));
+    gameVersion.setText(tableDetails.getGameVersion());
+    gameVersion.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setGameVersion(newValue));
 
     gameTypeCombo.setItems(FXCollections.observableList(Arrays.asList(GameType.values())));
-    GameType gt = manifest.getGameType();
-    if(gt != null) {
+    GameType gt = tableDetails.getGameType();
+    if (gt != null) {
       gameTypeCombo.valueProperty().setValue(gt);
     }
-    gameTypeCombo.valueProperty().addListener((observableValue, gameType, t1) -> manifest.setGameType(t1));
+    gameTypeCombo.valueProperty().addListener((observableValue, gameType, t1) -> tableDetails.setGameType(t1));
 
     gameYear.textProperty().addListener((observable, oldValue, newValue) -> {
       if (!newValue.matches("\\d*")) {
@@ -202,113 +369,253 @@ public class TableDataController implements Initializable, DialogController {
         gameYear.setText(s);
       }
     });
-    if (manifest.getGameYear() != null && manifest.getGameYear() > 0) {
-      gameYear.setText(String.valueOf(manifest.getGameYear()));
+    if (tableDetails.getGameYear() != null && tableDetails.getGameYear() > 0) {
+      gameYear.setText(String.valueOf(tableDetails.getGameYear()));
     }
     gameYear.textProperty().addListener((observable, oldValue, newValue) -> {
       if (newValue.length() > 0) {
-        manifest.setGameYear(Integer.parseInt(newValue));
+        tableDetails.setGameYear(Integer.parseInt(newValue));
       }
       else {
-        manifest.setGameYear(0);
+        tableDetails.setGameYear(0);
       }
     });
 
-    romName.setText(manifest.getRomName());
-    romName.textProperty().addListener((observable, oldValue, newValue) -> manifest.setRomName(newValue));
+    romName.setText(tableDetails.getRomName());
+    romName.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setRomName(newValue));
 
-    manufacturer.setText(manifest.getManufacturer());
-    manufacturer.textProperty().addListener((observable, oldValue, newValue) -> manifest.setManufacturer(newValue));
+    manufacturer.setText(tableDetails.getManufacturer());
+    manufacturer.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setManufacturer(newValue));
 
     SpinnerValueFactory.IntegerSpinnerValueFactory factory = new SpinnerValueFactory.IntegerSpinnerValueFactory(0, 4, 0);
     numberOfPlayers.setValueFactory(factory);
-    if (manifest.getNumberOfPlayers() != null) {
-      numberOfPlayers.getValueFactory().setValue(manifest.getNumberOfPlayers());
+    if (tableDetails.getNumberOfPlayers() != null) {
+      numberOfPlayers.getValueFactory().setValue(tableDetails.getNumberOfPlayers());
     }
-    numberOfPlayers.getValueFactory().valueProperty().addListener((observable, oldValue, newValue) -> manifest.setNumberOfPlayers(Integer.parseInt(String.valueOf(newValue))));
+    numberOfPlayers.getValueFactory().valueProperty().addListener((observable, oldValue, newValue) -> tableDetails.setNumberOfPlayers(Integer.parseInt(String.valueOf(newValue))));
 
-    tags.setText(manifest.getTags());
-    tags.textProperty().addListener((observable, oldValue, newValue) -> manifest.setTags(newValue));
+    tags.setText(tableDetails.getTags());
+    tags.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setTags(newValue));
 
-    category.setText(manifest.getCategory());
-    category.textProperty().addListener((observable, oldValue, newValue) -> manifest.setCategory(newValue));
+    category.setText(tableDetails.getCategory());
+    category.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setCategory(newValue));
 
-    author.setText(manifest.getAuthor());
-    author.textProperty().addListener((observable, oldValue, newValue) -> manifest.setAuthor(newValue));
+    author.setText(tableDetails.getAuthor());
+    author.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setAuthor(newValue));
 
-    launchCustomVar.setText(manifest.getLaunchCustomVar());
-    launchCustomVar.textProperty().addListener((observable, oldValue, newValue) -> manifest.setLaunchCustomVar(newValue));
+    launchCustomVar.setText(tableDetails.getLaunchCustomVar());
+    launchCustomVar.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setLaunchCustomVar(newValue));
 
-    keepDisplays.setText(manifest.getKeepDisplays());
-    keepDisplays.textProperty().addListener((observable, oldValue, newValue) -> manifest.setKeepDisplays(newValue));
+    keepDisplays.setText(tableDetails.getKeepDisplays());
+    keepDisplays.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setKeepDisplays(newValue));
 
     SpinnerValueFactory.IntegerSpinnerValueFactory ratingFactory = new SpinnerValueFactory.IntegerSpinnerValueFactory(0, 10, 0);
     gameRating.setValueFactory(ratingFactory);
-    if (manifest.getGameRating() != null) {
-      gameRating.getValueFactory().setValue(manifest.getGameRating());
+    if (tableDetails.getGameRating() != null) {
+      gameRating.getValueFactory().setValue(tableDetails.getGameRating());
     }
-    gameRating.getValueFactory().valueProperty().addListener((observable, oldValue, newValue) -> manifest.setGameRating(Integer.parseInt(String.valueOf(newValue))));
+    gameRating.getValueFactory().valueProperty().addListener((observable, oldValue, newValue) -> tableDetails.setGameRating(Integer.parseInt(String.valueOf(newValue))));
 
-    dof.setText(manifest.getDof());
-    dof.textProperty().addListener((observable, oldValue, newValue) -> manifest.setDof(newValue));
+    dof.setText(tableDetails.getDof());
+    dof.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setDof(newValue));
 
-    IPDBNum.setText(manifest.getIPDBNum());
-    IPDBNum.textProperty().addListener((observable, oldValue, newValue) -> manifest.setIPDBNum(newValue));
+    IPDBNum.setText(tableDetails.getIPDBNum());
+    IPDBNum.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setIPDBNum(newValue));
 
-    altRunMode.setText(manifest.getAltRunMode());
-    altRunMode.textProperty().addListener((observable, oldValue, newValue) -> manifest.setAltRunMode(newValue));
+    altRunMode.setText(tableDetails.getAltRunMode());
+    altRunMode.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setAltRunMode(newValue));
 
-    url.setText(manifest.getUrl());
-    url.textProperty().addListener((observable, oldValue, newValue) -> manifest.setUrl(newValue));
+    url.setText(tableDetails.getUrl());
+    url.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setUrl(newValue));
 
-    designedBy.setText(manifest.getDesignedBy());
-    designedBy.textProperty().addListener((observable, oldValue, newValue) -> manifest.setDesignedBy(newValue));
+    designedBy.setText(tableDetails.getDesignedBy());
+    designedBy.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setDesignedBy(newValue));
 
-    notes.setText(manifest.getNotes());
-    notes.textProperty().addListener((observable, oldValue, newValue) -> manifest.setNotes(newValue));
+    notes.setText(tableDetails.getNotes());
+    notes.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setNotes(newValue));
 
-    custom2.setText(manifest.getCustom2());
-    custom2.textProperty().addListener((observable, oldValue, newValue) -> manifest.setCustom2(newValue));
+    custom2.setText(tableDetails.getCustom2());
+    custom2.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setCustom2(newValue));
 
-    custom3.setText(manifest.getCustom3());
-    custom3.textProperty().addListener((observable, oldValue, newValue) -> manifest.setCustom3(newValue));
+    custom3.setText(tableDetails.getCustom3());
+    custom3.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setCustom3(newValue));
 
-    extrasTab.setDisable(manifest.getSqlVersion() < 64);
-    if(extrasTab.isDisable()) {
+    extrasTab.setDisable(tableDetails.getSqlVersion() < 64);
+
+    statusCombo.setItems(FXCollections.observableList(TABLE_STATUSES));
+    if (tableDetails.getSqlVersion() >= 64) {
+      statusCombo.setItems(FXCollections.observableList(TABLE_STATUSES_15));
+    }
+
+    if (tableDetails.getStatus() >= 0 && tableDetails.getStatus() <= 3) {
+      TableStatus tableStatus = TABLE_STATUSES_15.get(tableDetails.getStatus());
+      statusCombo.setValue(tableStatus);
+    }
+
+    statusCombo.valueProperty().addListener((observableValue, tableStatus, t1) -> {
+      this.tableDetails.setStatus(t1.getValue());
+    });
+
+    List<String> launcherList = new ArrayList<>(tableDetails.getLauncherList());
+    launcherList.add(0, null);
+    launcherCombo.setItems(FXCollections.observableList(launcherList));
+    launcherCombo.setValue(tableDetails.getAltLaunchExe());
+
+    launcherCombo.valueProperty().addListener((observableValue, s, t1) -> {
+      this.tableDetails.setAltLaunchExe(t1);
+    });
+
+    //displays
+    String keepDisplays = tableDetails.getKeepDisplays();
+    if (StringUtils.isEmpty(keepDisplays)) {
+      useEmuDefaultsCheckbox.setSelected(true);
+    }
+    else if (keepDisplays.equalsIgnoreCase("NONE")) {
+      hideAllCheckbox.setSelected(true);
+    }
+    else {
+      String[] split = keepDisplays.split(",");
+      for (String screen : split) {
+        if (StringUtils.isEmpty(screen)) {
+          continue;
+        }
+
+        int id = Integer.parseInt(screen);
+        switch (id) {
+          case 0: {
+            topperCheckbox.setSelected(true);
+            break;
+          }
+          case 1: {
+            dmdCheckbox.setSelected(true);
+            break;
+          }
+          case 2: {
+            backglassCheckbox.setSelected(true);
+            break;
+          }
+          case 3: {
+            playfieldCheckbox.setSelected(true);
+            break;
+          }
+          case 4: {
+            musicCheckbox.setSelected(true);
+            break;
+          }
+          case 5: {
+            apronCheckbox.setSelected(true);
+            break;
+          }
+          case 6: {
+            wheelbarCheckbox.setSelected(true);
+            break;
+          }
+          case 7: {
+            loadingCheckbox.setSelected(true);
+            break;
+          }
+          case 8: {
+            otherCheckbox.setSelected(true);
+            break;
+          }
+          case 9: {
+            flyerCheckbox.setSelected(true);
+            break;
+          }
+          case 10: {
+            helpCheckbox.setSelected(true);
+            break;
+          }
+        }
+      }
+    }
+
+    if (tableDetails.getVolume() != null) {
+      try {
+        volumeSlider.setValue(Integer.parseInt(tableDetails.getVolume()));
+      } catch (NumberFormatException e) {
+        LOG.error("Failed to set valume: " + e.getMessage());
+      }
+    }
+    else {
+      volumeSlider.setValue(100);
+    }
+    volumeSlider.valueProperty().addListener((observableValue, number, t1) -> tableDetails.setVolume(String.valueOf(t1.intValue())));
+
+    //cancel edit here!!!!!
+    if (extrasTab.isDisable()) {
       return;
     }
 
-    custom4.setText(manifest.getCustom4());
-    custom4.textProperty().addListener((observable, oldValue, newValue) -> manifest.setCustom4(newValue));
+    custom4.setText(tableDetails.getCustom4());
+    custom4.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setCustom4(newValue));
 
-    custom5.setText(manifest.getCustom5());
-    custom5.textProperty().addListener((observable, oldValue, newValue) -> manifest.setCustom5(newValue));
+    custom5.setText(tableDetails.getCustom5());
+    custom5.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setCustom5(newValue));
 
-    altRomName.setText(manifest.getRomAlt());
-    altRomName.textProperty().addListener((observable, oldValue, newValue) -> manifest.setRomAlt(newValue));
+    altRomName.setText(tableDetails.getRomAlt());
+    altRomName.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setRomAlt(newValue));
 
-    webDbId.setText(manifest.getWebGameId());
-    webDbId.textProperty().addListener((observable, oldValue, newValue) -> manifest.setWebGameId(newValue));
+    webDbId.setText(tableDetails.getWebGameId());
+    webDbId.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setWebGameId(newValue));
 
-    webLink.setText(manifest.getWebLink2Url());
-    webLink.textProperty().addListener((observable, oldValue, newValue) -> manifest.setWebLink2Url(newValue));
+    webLink.setText(tableDetails.getWebLink2Url());
+    webLink.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setWebLink2Url(newValue));
 
-    tourneyId.setText(manifest.getTourneyId());
-    tourneyId.textProperty().addListener((observable, oldValue, newValue) -> manifest.setTourneyId(newValue));
+    tourneyId.setText(tableDetails.getTourneyId());
+    tourneyId.textProperty().addListener((observable, oldValue, newValue) -> tableDetails.setTourneyId(newValue));
 
-    modCheckbox.setSelected(manifest.isMod());
-    modCheckbox.selectedProperty().addListener((observableValue, aBoolean, t1) -> manifest.setMod(t1));
+    modCheckbox.setSelected(tableDetails.isMod());
+    modCheckbox.selectedProperty().addListener((observableValue, aBoolean, t1) -> tableDetails.setMod(t1));
 
-    gDetails.setText(manifest.getgDetails());
-    gDetails.textProperty().addListener((observableValue, oldValue, newValue) -> manifest.setgDetails(newValue));
+    gDetails.setText(tableDetails.getgDetails());
+    gDetails.textProperty().addListener((observableValue, oldValue, newValue) -> tableDetails.setgDetails(newValue));
 
-    gLog.setText(manifest.getgLog());
-    gLog.textProperty().addListener((observableValue, oldValue, newValue) -> manifest.setgLog(newValue));
+    gLog.setText(tableDetails.getgLog());
+    gLog.textProperty().addListener((observableValue, oldValue, newValue) -> tableDetails.setgLog(newValue));
 
-    gPlayLog.setText(manifest.getgPlayLog());
-    gPlayLog.textProperty().addListener((observableValue, oldValue, newValue) -> manifest.setgPlayLog(newValue));
+    gPlayLog.setText(tableDetails.getgPlayLog());
+    gPlayLog.textProperty().addListener((observableValue, oldValue, newValue) -> tableDetails.setgPlayLog(newValue));
 
-    gNotes.setText(manifest.getgNotes());
-    gNotes.textProperty().addListener((observableValue, oldValue, newValue) -> manifest.setgNotes(newValue));
+    gNotes.setText(tableDetails.getgNotes());
+    gNotes.textProperty().addListener((observableValue, oldValue, newValue) -> tableDetails.setgNotes(newValue));
+  }
+
+  static class TableStatus {
+    private final int value;
+    private final String label;
+
+    TableStatus(int value, String label) {
+      this.value = value;
+      this.label = label;
+    }
+
+    public int getValue() {
+      return value;
+    }
+
+    public String getLabel() {
+      return label;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+      if (this == o) return true;
+      if (!(o instanceof TableStatus)) return false;
+
+      TableStatus that = (TableStatus) o;
+
+      return value == that.value;
+    }
+
+    @Override
+    public int hashCode() {
+      return value;
+    }
+
+    @Override
+    public String toString() {
+      return label;
+    }
   }
 }